Extract platform
