+++ src/types/liveness.rs
@@ -3,37 +3,7 @@ use std::{
     str::FromStr,
 };
 
-use radius_sdk::signature::ChainType;
-
 use crate::{error::Error, types::prelude::*};
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Deserialize, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum Platform {
-    Ethereum,
-    Local,
-}
-
-impl From<Platform> for ChainType {
-    fn from(value: Platform) -> Self {
-        match value {
-            Platform::Ethereum => ChainType::Ethereum,
-            Platform::Local => ChainType::Ethereum,
-        }
-    }
-}
-
-impl FromStr for Platform {
-    type Err = Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "ethereum" | "Ethereum" => Ok(Self::Ethereum),
-            "local" | "Local" => Ok(Self::Local),
-            _ => Err(Error::UnsupportedPlatform),
-        }
-    }
-}
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Deserialize, Serialize)]
 #[serde(rename_all = "snake_case")]
